+++ coreblocks/priv/traps/instr_counter.py
